Changed double.Parse() to int.Parse()

+++ Assignment2/Delegate.cs
@@ -6,5 +6,5 @@ public class Delegates{
 
     public static Func<double,double, double> ProductOfTwoDoubles = (double x, double y) => x * y;
 
-    public static Func<int,string, bool> NumericallyEqual = (int i, string k) => i == double.Parse(k);
+    public static Func<int,string, bool> NumericallyEqual = (int i, string k) => i == int.Parse(k);
 }
